adding locaiton to swiping sheets
