--- App/Components/Location/Locaiton.js
@@ -1,88 +1,0 @@
-'use strict';
-
-var React = require('react-native');
-var {
-  AppRegistry,
-  StyleSheet,
-  Text,
-  View,
-} = React;
-
-exports.title = 'Location';
-
-var Location = React.createClass({
-  watchID: (null: ?number),
-
-  getInitialState: function() {
-    return {
-      initialPosition: 'unknown',
-      lastPosition: 'unknown',
-    };
-  },
-
-  componentDidMount: function() {
-    navigator.geolocation.getCurrentPosition(
-      (initialPosition) => this.setState({initialPosition}),
-      (error) => alert(error.message),
-      {enableHighAccuracy: true, timeout: 20000, maximumAge: 1000}
-    );
-    this.watchID = navigator.geolocation.watchPosition((lastPosition) => {
-      this.setState({lastPosition});
-    });
-  },
-
-  componentWillUnmount: function() {
-    navigator.geolocation.clearWatch(this.watchID);
-  },
-
-  render: function() {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-          You were here
-        </Text>
-        <Text>
-          <Text style={styles.title}>Initial position: </Text>
-          {JSON.stringify(this.state.initialPosition.coords)}
-        </Text>
-        <Text style={styles.welcome}>
-          But now you're here
-        </Text>
-        <Text>
-          as at {JSON.stringify(this.state.lastPosition.timestamp)}
-        </Text>
-        <Text>
-          <Text style={styles.title}>Current position: </Text>
-          {JSON.stringify(this.state.lastPosition.coords)}
-        </Text>
-        <Text style={styles.welcome}>
-          Thanks science!
-        </Text>
-      </View>
-    );
-  }
-});
-
-var styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-  title: {
-  fontWeight: '500',
-  },
-});
-
-AppRegistry.registerComponent('constellation_native', () => Location);
